chore: pages polish
