+++ apps/build-onchain-apps/src/components/home/Guide.tsx
@@ -1,0 +1,12 @@
+import { Flex } from '@radix-ui/themes';
+import CodeBlock from '../CodeBlock';
+
+export default function Guide() {
+  return (
+    <Flex>
+      <CodeBlock>
+        <span>$</span> <span>npx @coinbase/build-onchain-apps@latest create</span>
+      </CodeBlock>
+    </Flex>
+  );
+}
