+++ apps/build-onchain-apps/src/components/home/Home.tsx
@@ -1,0 +1,24 @@
+import { Flex } from '@radix-ui/themes';
+import Guide from './Guide';
+
+export default function Home() {
+  return (
+    <>
+      <Flex direction="column" align="start">
+        <p className="my-2">
+          Build Onchain Apps takes an opinionated approach to streamlining and automating early
+          decisions you must make in building your consumer product.
+        </p>
+        <p className="my-2">
+          The onchain app generated by the CLI aims to be a companion that walks alongside you over
+          the first few months of building.
+        </p>
+        <p className="my-2">
+          So, if you are either a hackathon participant and/or an ambitious entrepreneur aiming to
+          establish the next successful company, this is built with you in mind. 💙
+        </p>
+      </Flex>
+      <Guide />
+    </>
+  );
+}
